+++ spec/himg_spec.rb
@@ -106,7 +106,7 @@ RSpec.describe Himg do
 
   it "accepts gpu option for GPU rendering" do
     if ENV['CI']
-      expect { Himg.render("<html></html>", gpu: true) }.to raise_error(/No compatible device found/)
+      expect { Himg.render("<html></html>", gpu: true) }.to raise_error(Himg::GpuNotFound, /No compatible device found/)
     else
       png_string = Himg.render("<html></html>", gpu: true)
       expect(png_string).to start_with("\x89PNG\r\n\x1A\n".b)